Rename to follow naming convention
